# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PublicProjectCategory = {
  name?: string;
} | null;

export interface PublicProject {
  id: string;
  title: string;
  short_description?: string | null;
  full_description?: string | null;
  target_amount: number;
  raised_amount: number;
  status: string;
  featured: boolean;
  category?: PublicProjectCategory;
}

export interface PublicMasjid {
  id: string;
  name: string;
  location: string;
  city: string;
  description?: string | null;
  image_url?: string | null;
  is_urgent: boolean;
  is_featured: boolean;
}

export interface HomeSlide {
  id: string;
  title: string;
  subtitle?: string | null;
  image_url?: string | null;
  action_type?: string | null;
  action_id?: string | null;
}