# lib/api-zod/src/generated/types/homeSlide.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HomeSlide {
  id: string;
  title: string;
  subtitle?: string | null;
  image_url?: string | null;
  action_type?: string | null;
  action_id?: string | null;
}