# lib/api-zod/src/generated/types/publicMasjid.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PublicMasjid {
  id: string;
  name: string;
  location: string;
  city: string;
  description?: string | null;
  image_url?: string | null;
  is_urgent: boolean;
  is_featured: boolean;
}