# lib/api-zod/src/generated/types/publicProject.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublicProjectCategory } from './publicProjectCategory';

export interface PublicProject {
  id: string;
  title: string;
  short_description?: string | null;
  full_description?: string | null;
  target_amount: number;
  raised_amount: number;
  status: string;
  featured: boolean;
  category?: PublicProjectCategory;
}